Add creator field in Event models and update post, update and delete events routes

+++ routes/events.py
@@ -71,6 +71,7 @@ async def retrieve_event(id: PydanticObjectId)-> Event:
 
 @event_router.post("/new")
 async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message" : "Event created successfully"
@@ -79,6 +80,12 @@ async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
 
 @event_router.put("/{id}", response_model=Event)
 async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code = status.HTTP_400_BAD_REQUEST,
+            detail = "Operation not allowed"
+        )
     update_event = await event_database.update(id, body)
     if not update_event:
         raise HTTPException(
@@ -90,6 +97,12 @@ async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depe
 
 @event_router.delete("/{id}")
 async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code = status.HTTP_404_NOT_FOUND,
+            detail = "Event not found"
+        )
     event = await event_database.delete(id)
     if not event: 
         raise HTTPException(
